requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Mail==0.9.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.0.2
zipp==3.18.1
